line_provider/pyproject.toml
[tool.poetry]
name = "line-provider"
version = "0.1.0"
description = ""
authors = ["ilia <boroznovskiyi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.4"}
httpx = "^0.27.2"
pydantic-settings = "^2.6.1"
redis = "^5.2.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.13.3"
orjson = "^3.10.11"
ruff = "^0.7.2"
pika = "^1.3.2"
aio-pika = "^9.5.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
exclude = ["migrations/*"]


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F811"]
